Refactor ValidationErrors.cs
